feat(swc): Enable swc behind a switch
