+++ .gitignore
@@ -2,4 +2,5 @@ node_modules
 pkg
 .envrc.local
 test/supervise/run-scratch.ts
+test/reload/run-scratch.ts
 .direnv
