+++ src/Options.ts
@@ -5,6 +5,7 @@ export interface RunOptions {
   terminalCommands: boolean;
   reloadOnChanges: boolean;
   supervise: boolean;
+  useSwc: boolean;
 }
 
 export interface ProjectConfig {
