+++ src/index.ts
@@ -7,6 +7,7 @@ import path from "path";
 import readline from "readline";
 import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
+import { EsBuildCompiler } from "./EsBuildCompiler";
 import { MiniServer } from "./mini-server";
 import { RunOptions } from "./Options";
 import { Project } from "./Project";
@@ -37,6 +38,11 @@ export const cli = async () => {
       type: "boolean",
       description: "Supervise and restart the process when it exits indefinitely",
       default: false,
+    })
+    .option("swc", {
+      type: "boolean",
+      description: "Use SWC instead of esbuild",
+      default: false,
     }).argv;
 
   return await esbuildDev({
@@ -44,6 +50,7 @@ export const cli = async () => {
     terminalCommands: args.commands,
     reloadOnChanges: args.watch,
     supervise: args.supervise,
+    useSwc: args.swc,
   });
 };
 
@@ -140,9 +147,8 @@ export const esbuildDev = async (options: RunOptions) => {
     serverSocketPath = path.join(workDir, "ipc.sock");
   }
 
-  const project = new Project(workspaceRoot, await projectConfig(findRoot(process.cwd())));
-  // project.compiler = new Compiler(workspaceRoot, workDir);
-  project.compiler = new SwcCompiler(workspaceRoot);
+  const compiler = options.useSwc ? new SwcCompiler(workspaceRoot) : new EsBuildCompiler(workspaceRoot, workDir);
+  const project = new Project(workspaceRoot, await projectConfig(findRoot(process.cwd())), compiler);
   project.supervisor = new Supervisor([...childProcessArgs(), ...options.argv], serverSocketPath, options, project);
 
   if (options.reloadOnChanges) startFilesystemWatcher(project);
