+++ test/oom/test.sh
@@ -2,6 +2,6 @@
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 set -ex
 
-$DIR/../../pkg/esbuild-dev.bin.js --max-old-space-size=50 $DIR/run.ts 2>&1 | grep "ReportOOMFailure"
+$DIR/../../pkg/esbuild-dev.bin.js $@ --max-old-space-size=50 $DIR/run.ts 2>&1 | grep "ReportOOMFailure"
 
 echo "found OOM failure"
