+++ test/simple/test.sh
@@ -2,4 +2,4 @@
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 set -ex
 
-$DIR/../../pkg/esbuild-dev.bin.js $DIR/run.ts | grep "IT WORKED"
+$DIR/../../pkg/esbuild-dev.bin.js $@ $DIR/run.ts | grep "IT WORKED"
