+++ test/test.sh
@@ -3,22 +3,26 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 echo $DIR
 set -e
 
-echo "::group::Simple test"
-$DIR/simple/test.sh
-echo "::endgroup::"
-echo
+ALL_ARGS=("--no-swc" "--swc")
 
-echo "::group::OOM test"
-$DIR/oom/test.sh 
-echo "::endgroup::"
-echo
+for args in "${ALL_ARGS[@]}"; do
+  echo "::group::Simple test ${args}"
+  $DIR/simple/test.sh $args
+  echo "::endgroup::"
+  echo
 
-echo "::group::Server test"
-bash $DIR/server/test.sh
-echo "::endgroup::"
-echo
+  echo "::group::OOM test ${args}"
+  $DIR/oom/test.sh $args
+  echo "::endgroup::"
+  echo
 
-echo "::group::Reload test"
-bash $DIR/reload/test.sh
-echo "::endgroup::"
-echo
+  echo "::group::Server test ${args}"
+  bash $DIR/server/test.sh $args
+  echo "::endgroup::"
+  echo
+
+  echo "::group::Reload test ${args}"
+  bash $DIR/reload/test.sh $args
+  echo "::endgroup::"
+  echo
+done
